backend/requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
contourpy==1.0.6
cycler==0.11.0
dj-database-url==1.0.0
Django==4.1.8
django-cors-headers==3.13.0
django-heroku==0.3.1
fonttools==4.38.0
gunicorn==20.1.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sqlparse==0.4.4
threadpoolctl==3.1.0
tzdata==2023.3
whitenoise==6.4.0
